Runs SPLASH at a single location for one day

Runs SPLASH at a single location for one day

run_one_day(lat, elv, n, y, wn, sf, tc, pn, kCw = 1.05, kWm = 150)

Arguments

lat
double, decimal degrees.

elv
double, elevation, m A.S.L. Default: \(0\).

n
double, day of year.

y
double, year. Default: \(0\).

wn
double, daily soil moisture content, mm (wn).

sf
double, fraction of sunshine hours. Default: \(1\).

tc
double, mean daily air temperature, degrees C. Default: \(23.0\).

pn
double, daily precipitation, mm/day.

kCw
double, supply constant, mm/hr. Default: \(1.05\) (Federer, 1982)

kWm
double, soil moisture capacity, mm. Default: \(150\) (Cramer-Prentice, 1988)

Value

List with the following components:

ho .......... daily solar irradiation, J/m2

hn .......... daily net radiation, J/m2

ppfd ........ daily PPFD, mol/m2

cond ........ daily condensation water, mm

eet ......... daily equilibrium ET, mm

pet ......... daily potential ET, mm

aet ......... daily actual ET, mm

wn .......... daily soil moisture, mm

ro .......... daily runoff, mm

References

Cramer, W. and Prentice, I.C., 1988. Simulation of regional soil moisture deficits on a European scale. Norsk Geografisk Tidsskrift - Norwegian Journal of Geography, 42(2-3), pp.149–151. doi: 10.1080/00291958808552193

Federer, C.A., 1982. Transpirational supply and demand: plant, soil, and atmospheric effects evaluated by simulation. Water Resources Research, 18(2), pp.355-362. doi: 10.1029/WR018i002p00355

Examples

soil <- run_one_day(lat = 37.7,
                    elv = 142,
                    n = 172,
                    y = 2000,
                    wn = 75,
                    sf = 1,
                    tc = 23,
                    pn = 5)

#> [1] "in run_one_day: pn = 5"

cat(sprintf("Soil moisture (run one day):\n"))

#> Soil moisture (run one day):

cat(sprintf("  Ho: %0.6f J/m2\n", soil$ho))

#>   Ho: 41646763.360234 J/m2

cat(sprintf("  Hn: %0.6f J/m2\n", soil$hn))

#>   Hn: 21774952.713925 J/m2

cat(sprintf("  PPFD: %0.6f mol/m2\n", soil$ppfd))

#>   PPFD: 62.042300 mol/m2

cat(sprintf("  EET: %0.6f mm/d\n", soil$eet))

#>   EET: 6.405468 mm/d

cat(sprintf("  PET: %0.6f mm/d\n", soil$pet))

#>   PET: 8.070889 mm/d

cat(sprintf("  AET: %0.6f mm/d\n", soil$aet))

#>   AET: 5.748034 mm/d

cat(sprintf("  Cn: %0.6f mm/d\n", soil$cond))

#>   Cn: 0.885192 mm/d

cat(sprintf("  Wn: %0.6f mm\n", soil$wn))

#>   Wn: 75.137158 mm

cat(sprintf("  RO: %0.6f mm\n", soil$ro))

#>   RO: 0.000000 mm