Calculate daily totals

Calculate daily totals updating the soil moisture until equilibrium.

spin_up(mdat, dtot)

Arguments

mdat
list with meteorological data (see the details section).

dtot
list with daily totals (see the details section).

Value

list, daily totals

Details

The list with meteorological data, mdat, should have the following components:

num_lines ..... double, length of meteorol. variable lists

lat_deg ....... double latitude (degrees)

elv_m ......... double, elevation (m)

year .......... double, year

sf ............ list, fraction of sunshine hours

tair .......... list, mean daily air temperature (deg. C)

pn ............ list, precipitation (mm/d)

The list with daily totals, dtot, should have the following component:

wm ............ list, daily soil moisture (mm)