Calculate daily evaporation fluxes

This function calculates daily radiation, condensation, and evaporation fluxes.

calc_daily_evap(
  lat,
  n,
  elv = 0,
  y = 0,
  sf = 1,
  tc = 23,
  sw = 1,
  ke = 0.0167,
  keps = 23.44,
  komega = 283,
  kw = 0.26
)

Arguments

lat
double, decimal degrees.

n
double, day of year.

elv
double, elevation, m A.S.L. Default: \(0\).

y
double, year. Default: \(0\).

sf
double, fraction of sunshine hours. Default: \(1\).

tc
double, mean daily air temperature, degrees C. Default: \(23.0\).

sw
double, evaporative supply rate, mm/hr. Default: \(1.0\).

ke
double, eccentricity of earth's orbit. Default: \(0.01670\), 2000CE (Berger, 1978).

keps
double, obliquity of earth's elliptic. Default: \(23.44\), 2000CE (Berger, 1978).

komega
double, lon. of perihelion, degrees Default: \(283\), 2000CE (Berger, 1978).

kw
double, PET entrainment, \((1 + kw) * EET\) Default: \(0.26\) (Priestley-Taylor, 1972)

Value

Returns a list object with the following variables:

nu_deg ............ true anomaly, degrees

lambda_deg ........ true longitude, degrees

dr ................ distance factor, unitless

delta_deg ......... declination angle, degrees

hs_deg ............ sunset angle, degrees

ra_j.m2 ........... daily extraterrestrial radiation, J/m^2

tau ............... atmospheric transmittivity, unitless

ppfd_mol.m2 ....... daily photosyn photon flux density, mol/m^2

hn_deg ............ net radiation hour angle, degrees

rn_j.m2 ........... daily net radiation, J/m^2

rnn_j.m2 .......... daily nighttime net radiation, J/m^2

econ_m3.j ......... water to energy conversion, m^3/J

cond_mm ........... daily condensation, mm

eet_mm ............ daily equilibrium evapotranspiration, mm

pet_mm ............ daily potential evapotranspiration, mm

hi_deg ............ intersection hour angle, degrees

aet_mm ............ daily actual evapotranspiration, mm

References

Berger, A.L., 1978. Long-term variations of daily insolation and Quaternary climatic changes. Journal of Atmospheric Sciences, 35(12), pp.2362-2367. doi:10.1175/1520-0469(1978)035<2362:LTVODI>2.0.CO;2

Priestley, C.H.B. and Taylor, R.J., 1972. On the assessment of surface heat flux and evaporation using large-scale parameters. Monthly weather review, 100(2), pp.81-92. doi:10.1175/1520-0493(1972)100<0081:OTAOSH>2.3.CO;2